requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
albumentations>=1.0.0
tqdm>=4.62.0
Pillow>=8.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
